Bug #883218: No documentation for installing OpenStack Compute with XenServer as hypervisor Bug #938408: Possible typo in a relevant technical term
Added <?dbhtml stop-chunking?> processing instruction so the introduction appears as a single page.
Add an introduction to Xen, XCP, and XenServer, with explanation of the
variety of technical terms.  This is included as part of the Installing
OpenStack Compute and Image Service chapter in the installation manual, and
the Hypervisors chapter of the compute admin manual.

Change-Id: I381e9b8ec19638ba90a8436fdcb6b5059f3ff963