Removes flag designation and --
In essex, flags are no more. They are configuration options.
This merge attempts to clean that up.
Also needed to assign xml:id to some sections in the Qpid write up - that ensures comments stay with the page they were intended for even if we rearrange
the content a bit.

Change-Id: I7d76b11347ca4b369646da74b36c6f79b3671a13