Make the publishing location a maven parameter.
The Jenkins jobs will be changed so that they are triggered by gerrit
ref-updated events, which provide information about the branch that has
just been updated.  We'll run a script at the beginning of the job that
determines where the content should be published based on the branch
(trunk, diablo, etc), and then pass that as a paremeter to the maven
build job.

The default is set to "unknown" so that an accidental build without
the parameter set will not overwrite something important (like trunk).

Jobs for books that are not yet core projects should manually set the
value to 'incubated', ignoring the value from the script:

 mvn -Drelease.path.name=incubated ...

comments.enabled is also added as a parameter to be set similarly.

Change-Id: I93f195f2bd1a73414a3995c0aff5835aae6501d6