Updates live migration doc to work with nova user
Fixes bug 887358
Update to Live Migration section of the compute manual to address
issues with the NFS configuration that would cause it to fail if
used with Diablo or Essex.

Where possible I have tried to refer readers to the excellent Ubuntu
documentation in preference for copy & pasting sections on generic
systems administration tasks (i.e. configuring NFS).

I have only tested this doc on a similar, but not identical setup,
so there could well be issues remaining. However, I feel that this is
an improvement enough (it has a chance of working!) to check it in.
Your feedback most welcome.

Fixes bug 891082
By linking to the existing GlusterFS document that can be used in
place of NFS

Cherry-picked from https://review.openstack.org/#/c/7602/

Change-Id: If076bf9103d71e584bdffc026741947f133fa381