Replace eucatools with references to nova CLI
fixes bug 891552

It has been decided to prefer novaclient over euca2ools when novaclient
does the job, so this fix is to bring the compute manual to implement
the remaining changes to bring it to fuitition.

Most of the work here is in the volumes section. Essentialy hacked
away to replace euca-*-volume commands with nova volume-*.

Also removed some old volumes-related text that seemed irrelevant in
current releases.

Change-Id: I020f86d337527c38af719703909ad52ae35f2d7b