Document underlying technologies
This is a first attempt to give the reader a sense of what the underlying
technologies are behind OpenStack Compute.

Shows two possible configurations, as well as a table with supported
technologies.

Comments welcome, especially if there are any inaccuracies or unclear
concepts. It was difficult to try to capture the interactions between
these technologies, especially on the networking side.

Cherry picked from https://review.openstack.org/8465 with ZeroMQ reference
removed since ZeroMQ is not supported in Essex.

Change-Id: Ib66e9c5b7c27b214b2b363d78f872ff5357b1619