Creates new disqus ID for each install guide - yum and apt.
This change should avoid the mind-bending associated with knowing
which distro a commenter is commenting about on a single page.

However it also means all old comment threads will not display on
a given page for the install guide.

I'll see what I can do to migrate threads but it does get tricky since
threads-per-page cannot be migrated to two separate pages.

Change-Id: Ifc768a7b0b04880af0e8086e9a913e3650cf4aa4