Reorganisation of Dashboard install chapter
In Horizon, databases are only used for session storage, for which
there are a few options. The install doc used memcache for the session
 cache, but also asked users to install a MySQL database, which was
unnecessary.

This change creates a new section, based on the horizon developer
docs which explains all of the options for session storage in
OpenStack Dashboard. It moves the redundant MySQL configuration to
that section, and references it in a note.

This change also attempts to improve section-leveling throughout the
chapter.

Patch set 2 fixes a typo (depooyment) and removes a <title> element
from the ch_installdashboard.xml as it caused a duplicate display of
the title text.

fixes bug 997347

Change-Id: I9d47adc06c48eac14dc1e9f50b1b0bb133bd3f8d