Document how to create qcow and raw images.
This is the beginning of a refactoring of the section on creating
images. This change describes how to create QCOW2 and raw images,
using CentOS as an example.

I would like to eventually remove the existing docs on creating a
raw image, since it's out of date, but I wanted to get this
replacement doc get in for review first.

Change-Id: Ic4564833e453ee34deafedb3f5621d74624aef20