Remove authtoken reference for glance-api-paste.ini
Essex branch had an erroneous reference to adding authtoken
to a paste pipeline. The authtoken is already there, in a
different pipeline, and it's just a matter of choosing the
correct pipeline rather than editing the file.

This was fixed in trunk, but not in essex. This change brings
the two in line.

Change-Id: I6dc6069088696c2f9a3985d120bf7021fd52376e