Remove deprecated boot=on kvm flag.
Stop passing boot=on when invoking kvm manually, as this parameter
has been deprecated and is ignored.

Cherry-picked from https://review.openstack.org/11211

Also, the arguments for manually booting a windows image with KVM were
different in the master and stable/essex branches. I backported all of the
command-line arguments since I assumed the master doc branch was more
up to date on this, but would help if somebody with more KVM knowledge
could sanity-check that.

Change-Id: Ifd7c4607f8101868ab99ac5c80b0c040983fc18b