doc swift service start issues with no ring
fixes bug 996944

As in the bug report, users following this guide exactly
will encounter a failed service start at this point in the
manual. This isn't a big problem - the ring files just need to be
copied first.

Rather than significantly changing the flow of the docs, this change
adds a simple warning so that first-time users aren't alarmed.

Change-Id: Ibaff34a14e7d82f674b6fe9b47923f7465d8825b