compute info shouldn't be in object guide and vv
fixes 1026738

As noted in the bug report, some object storage content was found
in the compute guide, and the inverse was also true.

This patch separates the support and troubleshooting section in
common into three parts: common, compute and object-storage. It then
updates the guides so they have the relevant include.

Patchset rebases against (fast-moving) master.

Change-Id: I1176efea28da3f7a3e01a3e026f1a8e0517d23ce