Clarify the use of allow_resize_to_same_host Flag
Being referenced in both:

- nova/compute/api.py
- nova/compute/manager.py

This flag is required for both nova-api (that invokes the
resize operation on the scheduler), and nova-compute
(which is invoked by the scheduler with prep_resize)

This makes me wonder if there's a more fundamental problem
with lack of good separation of responsibilities between
the parties involved, and whether the Flag should be needed
only in one place. However this goes beyond the scope of
this documentation effort.

Fix bug 1071931.

Patch also adds a note to the table where this flag is mentioned
so that people can see that it's set to false by default and
that it must be set in both places where nova-compute and
nova-api are run.

Rebase against master.

Change-Id: I731b28f16be80684748c9812256e89faf3897672