Add documentation for IBM Storwize and SVC
Add a documentation section on the IBM Storwize and SVC volume driver
to the OpenStack Compute Administration Manual.
Create a new sub section in Chapter 11 under "Volume drivers"
that describes how to configure the Storwize and SVC driver.

Minor edits to change "flags" to options and indicate these are
always found in nova.conf.

Change-Id: I202084c868715dff0521158cb73b0614be181b02