Add a section on co-locating services
fixes bug 988053

As in the bug request, people are often interested in running
services from different OpenStack projects on one piece of kit.

This patch adds a new section in common on co-locating services,
explaining the catches and things to look out for.
It's then linked from the compute install guide's architecture
section, and the install guide's assumption section.

Change-Id: I42a2843cb69fffd523a18143dc41e5275d9e5215