Add a note about identical GID/UID for live migrat
fixes bug 1080680

As noted in the bug report, if the GID and UID are not identical
between the servers, live migration fails badly and requires
some cleanup. This change adds a step to ensure users are aware
this needs to be done before creating the mount.

patch 2: adds a new section to the System Administration chapter
for the proceedure to recover from a UID/GID mismatch

patch 3: clarifies that the proceedure from patch 2 is run on
nova-compute hosts

Cherry picked from https://review.openstack.org/#/c/17281/

Change-Id: Id656b2bb62a97430999fc2c7ffd93e51a804a464