Update OS_PASSWORD to match keystone scripted inst
fixes bug 1083674

As noted in the bug report, there is a mismatch between the password
used in the keystone scripted install, and the rest of the documentation.

So if you used the scripted install, various steps will fail.

Since it's simpler to change the documentation and the choice of this
variable is arbitary - update the docs.

Patchset rebases against master.

Change-Id: I05e8268637e8572aadc8b274047b392d8c7aba75