Adds a storage concepts explanation
fixes bug 938365

There was a lack of storage concepts in the early parts of the
documentation. This adds a new section with a kick-start table
to allow first time readers to quickly understand the different
types of storage in an OpenStack installation.

Just a few edits -- double-check the ephemeral explanation as
I took it from the aboutcompute.xml file.

Change-Id: I447423f98bc3e36b6018179cdd81d8bb6d56b221