Remove ref to customizing api-paste.ini
Customizing api-paste.ini with keystone auth info is no longer
required. This was mostly fixed in a previous doc patch[1], but there
was a residual statement about changing api-paste.ini that's no longer
relevant.

This patch removes that statement. We no longer show the example
api-paste.ini file in the "Configuring OpenStack Compute" section,
since no changes to the file are needed.

[1]: https://review.openstack.org/15310

Change-Id: I7e7a245d6ea0e7371c0fe6c79898bd5d7fbaf89e