Include rootwrap.xml in the compute admin guide
There was documentation in the repo that describes how rootwrap works,
but it wasn't actually included in any of the OpenStack doc books as
far as I could tell. This patch includes it in the compute admin guide,
which is where the file is located.

Cherry-picked from https://review.openstack.org/20551

Change-Id: I5e5e0581719e965f0949494623ec4b1de1817348