Fix XML schema validation failures
With this patch, running the tools/validate.py script will return
no DocBook RelaxNG schema validation failures.

Once this patch lands, and the equivalent lands in stable/folsom,
I will follow up on getting this run automatically by Jenkins so
we can see the results when there are new doc patch proposals.

There were three XML validation failures that were addressed

openstack-compute-admin/rootwrap.xml:

rootwrap.xml had to xml:id values that were the same. These have
been changed so they are different. It turns out this file was not
being included in the documentation at all. I submitted a separate
patch to include it: https://review.openstack.org/20551

openstack-compute-admin/computeconfigure.xml:

computeconfigure.xml had xi:include elements with an 'os' attribute.
This violates the spec, and it doesn't work, either. We had two
lines that looked like this:

<xi:include os="ubuntu" ...>
<xi:include os="rhel;fedora;centos" ...>

and both files were being included in the docs. This was changed so
that only one file is included, and there's a note that has been
added that describes the difference between the two files, which
is very minor (just one parameter).

openstack-ha/ha-guide-docinfo.xml:

This isn't a real XML file, it's a fragment of XML that is used
by a non-standard toolchain. The validation script has been configured
to just ignore it for now.

Change-Id: Ie77218efc56c16b7bf162f5e02fd7a6d99116dc1