Improves LVM compatibility with virtual machines
The previous solution of adding a volume filter prevents the issue described in bug #1090595, but lvm on the cinder server is still open to issues due to conflicting volume group names.
If multiple virtual machines create volume groups with the same name, lvm will become locked due to the conflict.

This prevents the issue by restricting LVM to only scan physical volumes that are explicitly stated.

Improves the fix for: bug #1090595

Change-Id: Ide66c18cfa65c8a3cc5a69940f59b863cf0ee5fc