Grizzly release prep - old releases, bugs deleted
This patch is the result of searching the documentation for
references to old releases, fixed bugs et al. It attempts to
remove and update information that will be obsolete when these
docs are released for Grizzly.

patch2 removes references to deprecated versions of ubuntu (11.10
 being the most recent)

Rebased after merging in changes for Grizzly to the install guide.

patch4 addresses Anne's comments

Rebased again, ready to merge.

Change-Id: Id57808802eed66145aa683240000f4f3e90706d0