Added documentation about GlusterFS volume support.
A small mention of Gluster was added to getstart.xml. A full description
of its options was added to the Block Storage guide. Finally, a table
was modified in the Ops guide to show it now has volume support.

Patchset adds more explanation about sparsed volumes settings.
Merge with master.

Change-Id: I36cb344e63f1bddf4125084e02ac507f56591a11
Fixes: bug #1117767