Add acknowledgements to compute-adminguide
The bug mentioned is above removing a by-line. It turns out that
the section the by-line applies to has been updated quite a bit from
the original. So, the proposal is to remove it and add a mention
in an acknowledgements section at the end of the document.

Feedback very welcome.

fixes bug 1035675

Change-Id: I392d1c0b74fc2cb9c647e47fed44186d1faf355b