improve VMware Nova docs, including most new grizzly items.
bug #1106404
bug #1100147

The existing VMware Nova docs were very outdated, incomplete,
and in a few cases, just plan wrong.  They also did not cover the new
VMwareVCDriver and associated updates that arrived in Grizzly.  This
commit removes a lot of stale content (and even graphics) and adds the
basic instructions for both the VMwareVCDriver and the VMwareESXDriver.

I filed an additional bug for some topics (disk details, vnc) that
I did not already cover (bug #1175000)

Change-Id: Id6a4c9022e07f82622aa40eedd0449956f0475e8