Bug # 1156429 - Correct placement of auth=no setting for Qpid.
Removed the auth="no" setting from the section of the guide it was in
which was unrelated to messaging and assumed that Qpid would be
colocated with Nova (hence the config file being updated didn't exist
for some users).

Updated the message broker installation section to more thoroughly cover
Qpid installation and set up and added missing rpc_backend directive to
the example nova.conf file for the "yum" version of the book. Also fixed
an errant prompt character in the installing MySQL material (s/$/#/)
because the introduction tells the user that the commands within the
section need to be run as root.

Change-Id: I89df0d1050c6871254001e560a45ea6189bbfe18