Add quantum configuration reference tables
This is the first attempt at automatically generated configuration
tables being added to documentation.
blueprint autogenerate-config-tables

https://wiki.openstack.org/wiki/Blueprint-autogenerate-config-tables

Work is being conducted out-of-repo at :

https://github.com/fifieldt/autogenerate-config-docs

quantum has existing tables, which are out of date for Grizzly.
These new tables will need to be linked in, but first submitting these for
review without linking them in to a book yet, because feedback is needed.

Please provide feedback on:
* completeness (is every option and its description present)
* groupings (is every option in the right group)
* potential integrations (can you see where one or more of these
tables could be combined under a nice descriptive paragraph)
* relationships to config files (are these useful as is, or how to
best integrate them with examples)
* anything else you can think of

Change-Id: I396b04edb668c0a5a403ccfd3f6db90c78d29ce6