Quantum: under the hood details
Add a chapter in Network Admin guide that describes how the Open
vSwitch plugin and Linux bridging plugins configure the various
network devices on the compute and networking hosts.

All of the images and code are from Etsuji Nakai (Red Hat) from the
slides:

http://www.slideshare.net/enakai/how-quantum-configures-virtual-networks-under-the-hood

This content has been added to the docs with Etsuji's permission, and
he has been recorded as author of this commit.

Lorin added some supporting text and specified (guessed) at the config
file options.

Change-Id: Ie2ce069b025bde273c20c8324317af24ddd4d019