Bug 1099573 Add Quantum metadata proxy documentation
Document metadata_agent.ini - includes auth_* and admin_* flags needed for
the Qauntum client as the metadata agent does not seem to be able to read
these from quantum.conf even though it is on the command line.
Removed section in advanced config options about the need for additional
routing as this no longer applies.

Change-Id: I2f5b73d5a6bd2add27f9dd46ad2ca1d6db832b1c