Refactored active/passive section of HA guide
As part of the restructuring of the HA guide to include
active/active, the original active/passive page files
are now renamed to be prefaced with "ap-" to distinguish
them from the active/active pages, prefaced with "aa-".
The main ha-guide.txt file has also been edited to
reflect this change.

bug 1196096
blueprint improve-high-availability-support

Change-Id: I303f502a641278d1cdc9663381afd4c2be7cc67c