Merge upload and verification of images sections
ch_images-uploading.xml and images-verifying-install.xml have duplicated
but out of sync informations. Merge details from ch_images-uploading.xml
in images-verifying-install.xml and remove the file from the book.

Also update the Image Service commands output to match the name and ID
used in the verifying section in the rest of the doc.

Change-Id: I60537f215617b7efd61af7aa87f1fa55da6f12d2