Bug # 1200232 - force_dhcp_release now defaults to True.
Historically it was was False because it requires the dnsmasq binary
'dhcp_release' which was not present in all distributions. Most
content and sample files already recommended setting the value to
True, I have updated the few instances I found where it seemed
applicable to highlight that True is now the default. I did leave one
mention where it is recommended to set False for RHEL 6.2 because this
is still the case.

There is probably a wider discussion needed there about whether
at this point we should recommend >= RHEL 6.3 or even 6.4 for
Grizzly/Havana as it significantly improves the availability of
packages for a number of issues like this).

Change-Id: Ief819e1d3366ee1e1cd42f6bd8b61eaf7c20401d
Fixes: bug 1200232