updated installation notes for SLES
Removed note about deprecated packages provided by B1 Systems
on the openSUSE Build Service. B1 Systems is now working on
the packages provided in the project Cloud:Master and it's
subprojects.

Updated the note about declaring the repository for SLES 11, the
lastes release of SLES 11 is SP3.

Added the note that there are no known issues with the Python
interpreter on SLES11 SP3.

Added the note that the latest development snapshots are available
in the project Cloud:OpenStack:Master on the openSUSE Build
Service.

Change-Id: I9408afdbd4b90495e2f34f5acb97a98dd5e7d179