NetApp documentation for Havana
Change updates the NetApp Unified driver documentation for Havana.

As the window to update the code's option text is closed for release,
this patch manually updates the automatically generated table
cinder-netapp.xml to allow for presenting the best possible text to
users on release. The code should be updated at a later date.

Change-Id: I7b3ffd11a5b60447d54449a7253c2f6bd2951c6e