Extend, structure Keystone troubleshooting section
This patch creates a new section in common for identity
troubleshooting, which supercedes the existing stub sections
in common and openstack install and adds PKI middleware
troubleshooting information.

The section is included in the basic install guide
(install identity chapter) and compute admin manual
(via the common identity chapter that is only used in this guide).

fixes bug 1204640

Conflicts:
	doc/src/docbkx/common/ch_identity_mgmt.xml
	doc/src/docbkx/common/section_identity_troubleshooting.xml
	doc/src/docbkx/openstack-install/identity-troubleshooting.xml

Change-Id: Ibce5d404d1a5d1c413f1324c897eca4b36f6edf2
bp:restructure-documentation