do not exit after first invalid file
At the moment tools/validate.py exits after it found an
invalid file.

I think it's better to always validate all files and to
exit then if any invalid file was found. This way the
creator of a review request can fix all found issues in
all files and has not to create several patch sets.

Change-Id: Ic5668f0791680802b8cef3b5e1016c9dc76479b9