validate modified files and build affected books
At the moment the validation script simply validates all
files and builds all books afterwards. This results in a
very long runtime (appr. 30 minutes at the moment).

With this patch only modified files are validated and only
affected books are built. A book is affected if one XML
file includes a modified file. If no book is directly
affected all books are built.

In this way, resources can be saved and validation is
performed much faster.

Change-Id: I17dcf119c0bafc4656c9f267f0d3f5ae884ea76e