Change whitespace check to be info only
The recent introduction of the whitespace check has been good for
improving our consistency, however while discussion is ongoing
about how to best manage whitespace, the check should be non-voting.

This patch (intended to be temporary while discussion continues)
changes the script to print when whitespace is wrong,
rather than leaving a -1 vote.

Change-Id: I082aa4418864cd50d087dfe01c3bfff4d6df9f8f