take care of books with sources in subdirectories
At the moment the validation doesn't take care of books
where the pom.xml is in an other directory than the sources
of the book. This patch will fix this issue.

For example: the pom.xml of basic-install is available at
doc/src/docbkx/basic-install/pom.xml but the sources of the
books are in doc/src/docbkx/basic-install/src. Because
the path to the pom.xml differs from the path of the sources
the validation will not check the files in the sources
directory for xi:include elements.

Change-Id: I3b92c3c4bff46ba6f9ac53b3ec56d24e7caf6ff6