Swift work for config reference
This patch re-works swift content for the new configuration reference
structure.

The Object Storage configuration section is pared down, with
additional/optional features moved to a new file and section of the
document - adding to what was previously there in the Object Storage
Admin Guide.

New stub sections added for drive-audit, static web and form post.

One of the big changes is the move of all option tables to
common/tables.

These have been generated with a new script which converts swift
 conf files into docbook tables, and attempts to find descriptions
of those configuration options in swift's devref documentation.

Future work for this script: merged in with the autodoc tool
structure, add an 'update' mode.

Change-Id: I75a3c442e4734594bec202fecbe4a4272d297ac7
lp:restructure-documentation