Remove unused file openstack-compute-admin/section_keystone_middleware.xml
This file was introduced as part of change
I6c35b2a50be25a3aa06e3f951dec8598315503ec. Looking at the history of the
change, the file was first split out from ch_identity_mgmt.xml but later
the contents moved back in - and the extra file was kept. So, this was committed by
accident and the file is unused.

Change-Id: I67da2bb17c5d79c4cb436f359a1c28cc105fd236