Improve Debian support in Basic Install Guide
The Guide is built for both Debian and Ubuntu and there are a few
places that are only relevant for either of these. These sections
have been updated to clarify for which OS they are.

Some places used "deb" instead of "debian" or missed to add os="debian",
these were fixed as well.

Note: The basic install guide is build with os="ubuntu;debian", so some
of these changes do not change the generated code.

Change-Id: Ib50c5cee239c370a17d5462c2a6837357242425d
Closes-Bug: #1206328