Add Docker to Hypervisors section
Docker is an open-source engine which automates the deployment of
applications as highly portable, self-sufficient containers which
are independent of hardware, language, framework, packaging system
and hosting provider.

Support for Docker has been added in Havana.

This patch adds the a section on Docker to the Hypervisors
section of the Configuration Reference. It requires the config table
found in this patch: https://review.openstack.org/#/c/45922/
before merging.

Change-Id: If89225036cccad9cc2c8eb3ce7cc3bf6bd710cf0
Closes-Bug: 1221207