Change all external subnet examples to have dhcp disabled
Make consistent use of --disable-dhcp in subnet-create console
examples for external networks. These subnets are used for floating
IP allocation pools and the DHCP attribute does not make sense.
There doesn't seem to be a problem when it is set, but the examples
in the doc might confuse readers if it is.

Bug: 1224992
Change-Id: I09f4cd0e4ab182e491b40aece55119deb99005bb