Avoid having twice Appendix A
Both user-guide and user-guide-admin had two versions appendices
labelled "A". The problem was setting appendixAutolabel to 1 in
the pom.xml file and manual overrides of the number in the files.
The patch cleanes all of this up and renames ch_resources.xml
to an appendix part.

security-guide and training-guide also set this variable but did
not have an appendix yet, so remove the variable there as well.

Change-Id: I446f4968b675d3286f2302071ddfa8037425fd6f