doc xen support for auto_disk_config=disabled
Previously, only True and False were supported. Havana changed to
allows an administrator to specify which images should reject
any user requests for auto config.

Also fix an incorrect statement that auto_disk_config was a nova.conf
variable - it is instead a property set on an image in glance.

Change-Id: I32a09ce1258893388d31ca2c1704e57d23ceff97
Closes-Bug: 1214816