Add sample architecture, fix ubuntu instructions
1) The install guide was lacking a sample architecture section

This adds one and includes an image, with a followup image
in the basic networking section.

2) The instructions for disabling NetworkManager were given for
all distributions. This is not necessary on Ubuntu Server or Suse
server

3) the network configuration samples were only for fedora
this adds one for ubuntu, and a placeholder for SUSE

4) you need to restart the network after making changes

5) ubuntu sets services to start on install, no need to
restart or chkconfig (which doesn't exist on ubuntu)

6) ubuntu does not have openstack-config or openstack-db

This install has now been verified for ubuntu from the beginning
to the end of the identity service install.

Change-Id: I4e41092ab3ae399c83bb01d491847b6ca1998cfa
Closes-Bug: 1237863
Closes-Bug: 1237890
Closes-Bug: 1237898
Closes-Bug: 1237906
Partial-Bug: 1236952