Use consistent RabbitMQ settings across guide.
Use consistent values for rabbit_host, rabbit_userid, and
rabbit_password across the guide. Previously there were a number of
variations, particularly of the host and password values.

In a subsequent patch I intend to cover off the fact that RabbitMQ
is installed/configured in at least two locations. I may depending
on feedback also end up adding an instruction to set the guest
user's password in which case a further update to rabbit_password
will be required.

I think just getting the initial values to base this work on consistent,
as I am proposing in this patch, is a good start.

Where programlisting elements that were touched use "bash" as the
language I have also flipped them to "ini".

Change-Id: Ie3779db58f8a397753a9a2409bd838ea97a87568