Add warning explaining sf_user_prefix option
There's an issue with HA environments and the default
config options used for the SolidFire driver.  The SolidFire
driver creates a unique account on the cluster for each tenant,
by default this account name is : <$hostname>-<$tenant-id>

This works great, unless you're in an HA env, in which case there are
two possible accounts that are valid.  Also in the case of upgrading
or migrating your cinder-volume service to a new node, the result is that
the formation of the account name is now for the "new" hostname and users
loose access to their existing volumes that they created on the "old" host.

Change-Id: Id2ffc268e59810decc85f8da4ec41e4b751d65ff