Fix Misplaced Steps in OVS Setup
It would appear that some of the steps from the main Neutron setup
managed to find themselves in the OVS section.  This was rather
confusing to the reader, as the steps seemed to abruptly change
course in the OVS documentation, while the steps in the main Neutron
install guide seem to skip over discussing the plugin entirely.
This is now rectified.

Change-Id: I022f2a9061229f067057c1e386df0ca1f8942724
Fixes: bug #1243749
backport: havana