Fix Misplaced Steps in OVS Setup
It would appear that some of the steps from the main Neutron setup
managed to find themselves in the OVS section.  This was rather
confusing to the reader, as the steps seemed to abruptly change
course in the OVS documentation, while the steps in the main Neutron
install guide seem to skip over discussing the plugin entirely.
This is now rectified.

Fixes: bug #1243749
backport: havana
(cherry picked from commit 0c114976eadce0c30ed3961c2d19c56a00fc4cc2)

Change-Id: Ic933b4ef3a91210aa13c9e031e0c61cc1e30ac77