Install Guide: Clairify VLAN vs GRE Differences
Previously, there was a rather cryptic note about how
GRE tunneling was easier to set up but "less flexible",
and that was about it with regards to the choice between
GRE and VLAN networking in Neutron.  Now, details about how
GRE is easier to set up because it doesn't require special
hardware configuration has been added, with a note about
how it uses its own little protocol, so filtering based on
contents on the physical network is harder.  Similarly,
it is now explained that VLANs, on the other hand, just
tack on to the ethernet header, making it easier to see
the contents but requiring special network hardware configuration
so as not to strip them off or overlap.

Note: This is the first in a series of commits designed to clean
      and clairify some points in the Neutron section

cherry-picked: https://review.openstack.org/#/c/53928/
backport: havana
Change-Id: Idac013b24ccdba0b8261f59a635831e92887c119