Install Guide: Fix Typos in Neutron Network Node
There were a myriad of small typos (such as missing 'the's, missing
replaceable tags, slightly incorrect wording, and a step that was
copied twice). These have been corrected, and working has been
modified slightly for clairity in several places.

This is the second in the series of clean-up commits

backport: havana

Change-Id: I322846ea66de7259966dd8253d0f35eca81c4928