Move hypervisor sections out of common
The purpose for common is to host files that are used in multiple guides.

These files were only in use in the configuration reference.

In addition, the content and structure of these files was not very
conducive to sharing between guides.

Therefore, this patch simply moves the hypervisor sections from common
 into the compute area of the config-reference.

Should (in the future) some content from these be required in -
 eg Install Guide or Cloud Admin Guide, it is likely best that they are
chopped apart with appropriate sections divided appropriately.

backport: havana

Change-Id: Ie153ccf810de2dccd32cb7b6ee3076fbac365605