Specifics for nova install in Debian.
This patch removes things which are not needed to be in
the Debian installation chapter for Nova, and adds some
Debian specifics (plus remove the "Notes for Debian
users).

Note that this is a break-down of:
https://review.openstack.org/54394/

into smaller patches.

backport: havana

Change-Id: I3c0f538acb392194733293478fa35e9824328fff
(cherry picked from commit 48a63add358e64e2bea79f59d43c245063b6def5)