Specifics for Debian install of Cinder
This patch reviews the installation procedure for Cinder
in the Debian install-guide. It also moves the part which
explains how to setup the volume group on top of the
procedure, since that's ok for all operating systems, and
that's needed for Debian (a volume group must be present
before installing cinder-volume, otherwise it will not
be detected automatically, which adds more manual work).

Change-Id: I84914a3447664f70e0c5678f0445b9385ae3b8c8
backport: havana