Debian consitency and links
This patch reviews the "Answer to the debconf" things in
every service, so that they look the same.

It also adds links to the debconf chapter, so that users
can navigate more easily in the doc.

It also changes the titles of the debconf chapter so that
they match the text in each chapter.

It also documents the fact that python-argparse should be
installed manually on all Wheezy setups, which is an issue
that has been reported by some users, and which wasn't
documented.

The patch also extends the explanations about the
meta-packages.

This patch adds a lot of <informalfigure> tags that were
missing as well.

Finally, the patch also moves the explanation about remote
MySQL setup from the nova installation procedure to the
debconf-dbconfig-common section of the debconf chapter,
which is a much better fit.

This is my last clean-up work, I believe.

Change-Id: I5a85f39e49786ae90cc0afbf13b94baf486408f4
backport: havana
(cherry picked from commit f174fc2caab209061069621bfe26fb3edb75d6c7)