Ensure all "os" values are valid/recognized
Added a check to ensure that all "os" values are valid/recognized. The
list of recognized values is contained in a global and can be updated
if/when we add any further operating systems.

Note that "adminuser" and "enduser" values are also included as
recognized as they are required by the user guides at this time.

Change-Id: I3303b9cb880b70cbf8298eecafc56dffe2276476