Install Guide: add a chapter about client commands
This new chapter is taken from the common/ section, and explains how to
install clients on machines outside the OpenStack cluster. The install
guide can be misleading about how to use clients to administrate the
cloud, adding this section will hopefully help clarifying this.

pom.xml is modified to add an audience target (installer), used to build
specific parts of the common files.

backport: none
Change-Id: Ib4187d86bb2e3d7d9a26df426e17b4f4039a90fa