restructure of associate training guide
cleaning up the work that was starting at the icehouse summit. Removed some
of the work we started as it was going in the wrong direction. Added in
timeline which helped to limit the amount of content available for a two
day class. operator guide will take on the rest of the material that was
in the associate guide.

backport:none
implements bp/training-manuals

Change-Id: I6656afddee16c3deeb61c1ebc91522f45fd6d300