Remove confusing message about mysql credentials
The Glance installation material contained a note indicating that if
connecting to MySQL fails the user should use the IP address of the
database server instead when specifying the credentials.

Besides being confusing in its wording this note referred to a situation
that the user could not possibly become aware of until starting the
Glance services later in the chapter.

Additionally no such warning is included elsewhere in the guide as its
presumed (via specification in the prerequisites up front) that the user
has set up their /etc/hosts files such that the controller host name will
resolve.

Change-Id: Ic5e6e1820628b0854effe7594b09610d030d0573
Closes-Bug: #1251458
backport:havana