rst-xml-conversion revisited
we likely will be using most of the developer reference material in
the developer training guide. there is good tidbits of information like
what is patched here. as it gets included, some reformating will get pushed
back into the cinder project. like the architecture asc art will get either
converted into a diagram or add a flag for an xml programlisting tag. rst files
in the devref source, but not converted to xml will get patched in cinder
for update or removal.

implements blueprint rst-xml-conversion

Change-Id: Ib4a577a77d0d19e1864860c6940a46cbb11b2028