Grant access to 'dash'@'localhost' in dashboard setup
MySQL does not include local socket connection in the '%' wildcard. This
meant that following the existing instructions if the dash was installed
on the same system as the database it was not actually able to access
it.

A grant for 'dash'@'localhost' has been added to address this. It is
worth noting that in this situation (dash + MySQL on same host) this
means that the wildcard grant is actually insecure and not required, but
this is a wider issue that we need to address/highlight  at some point.

Currently we issue both grants for all services, this fix just brings
the dashboard instructions into line with the rest.

Change-Id: Id32432c3612945d6cc8860e947d332d15a8072b4
Closes-Bug: #1255975
backport: havana
(cherry picked from commit 1cad5969600f53f46b948a9ba6f6cd907f033a9c)