Clarify the step in image service install guide
When follows the instructions at "Install the Image Service in
OpenStack Installation Guide for Ubuntu 12.04 (LTS)  - havana".

In step 8, we have to add options to filter:authtoken section.
It is not clear though in glance-api-paste.ini what the contents
of the section should be, since it contains the option:
delay_auth_decision = true

Add some information in step 8 so it is not confusing any more.

Change-Id: Ib29c2138b115720e72bee5a13066c79158e188f7
backport: havana
Closes-Bug: #1255552